refactor(web): decompose complex views and resolve relative import lint errors
- Refactored Medicines.jsx, Settings.jsx and Dashboard.jsx sub-components.
- Extracted business logic into feature-specific hooks.
- Configured @settings path alias in Vite and ESLint.
- Moved medicines sub-components to medications feature domain.
- Strictly enforced absolute path aliases to resolve quality gate violations.
- Verified 0-error state via validate:agent.
