+++ apps/web/src/features/protocols/components/redesign/ProtocolRow.jsx
@@ -1,28 +1,6 @@
-/**
- * ProtocolRow — Row reutilizável com dois variantes
- *
- * variant="card" (padrão — Modo Simples):
- * - Layout vertical (flex-column) em mobile
- * - Cards empilhados com nome, dosagem, frequência, adesão, estoque
- * - Expandível para titulação e notas
- *
- * variant="tabular" (Modo Complexo Desktop):
- * - 4 células separadas: nome | frequência | adesão | estoque
- * - Renderiza sem button, sem expand (cada célula é independente)
- * - CSS Grid do container pai distribui as 4 células horizontalmente
- */
-import { Trash2 } from 'lucide-react'
-import { AnimatePresence, motion } from 'framer-motion'
-import AdherenceBar7d from './AdherenceBar7d'
-import AdherenceLabel from './AdherenceLabel'
-import StockPill from './StockPill'
-import TitrationBadge from './TitrationBadge'
+import ProtocolRowTabular from './ProtocolRowTabular'
+import ProtocolRowCard from './ProtocolRowCard'
 
-/**
- * ProtocolRow — Row reutilizável com dois variantes
- * S7.5.5: Adicionar isHovered prop para iluminar linha inteira em desktop
- * Célula 1 convertida de button para div (clique agora no wrapper pai)
- */
 export default function ProtocolRow({
   item,
   isComplex,
@@ -32,177 +10,36 @@ export default function ProtocolRow({
   onDelete,
   activeTab,
   variant = 'card',
-  isHovered = false, // S7.5.5
-  onRowMouseEnter, // S7.5.5
-  onRowMouseLeave, // S7.5.5
-  onRowClick, // S7.5.5
+  isHovered = false,
+  onRowMouseEnter,
+  onRowMouseLeave,
+  onRowClick,
 }) {
   const showAdherence = activeTab === 'ativos'
 
-  // ────────────────────────────────────────────────────────────────
-  // VARIANTE: TABULAR (Modo Complexo Desktop)
-  // ────────────────────────────────────────────────────────────────
   if (variant === 'tabular') {
-    const hoverClass = isHovered ? 'protocol-row-tabular__cell--hovered' : '' // S7.5.5
-
     return (
-      <>
-        {/* CÉLULA 1: Nome + Dosagem — S7.5.5: recebe mouse events e click */}
-        <div
-          className={`protocol-row-tabular__cell protocol-row-tabular__name-cell ${hoverClass}`}
-          style={{ minHeight: '3.5rem' }}
-          onMouseEnter={onRowMouseEnter}
-          onMouseLeave={onRowMouseLeave}
-          role="gridcell"
-        >
-          <button
-            type="button"
-            className="protocol-row-tabular__name-btn"
-            onClick={onRowClick}
-            onKeyDown={(e) => {
-              if ((e.key === 'Enter' || e.key === ' ') && onRowClick) {
-                e.preventDefault()
-                onRowClick()
-              }
-            }}
-          >
-            <div className="protocol-row-tabular__name-row">
-              <span className="medicine-name">{item.medicineName}</span>
-              {item.concentrationLabel && (
-                <span className="protocol-row__dosage">{item.concentrationLabel}</span>
-              )}
-            </div>
-            <div className="protocol-row-tabular__intake">{item.intakeLabel}</div>
-          </button>
-
-          {onDelete && isHovered && (
-            <button
-              className="protocol-row-tabular__delete-btn"
-              onClick={(e) => {
-                e.stopPropagation()
-                onDelete(item)
-              }}
-              aria-label={`Excluir tratamento ${item.medicineName}`}
-              title="Excluir tratamento"
-            >
-              <Trash2 size={14} />
-            </button>
-          )}
-        </div>
-
-        {/* CÉLULA 2: Frequência + Horários — S7.5.5: aplicar hover class */}
-        <div
-          className={`protocol-row-tabular__cell protocol-row-tabular__schedule-cell ${hoverClass}`}
-        >
-          <div className="protocol-row-tabular__frequency">{item.frequencyLabel}</div>
-          {item.timeSchedule.length > 0 && (
-            <div className="protocol-row-tabular__times">{item.timeSchedule.join(' / ')}</div>
-          )}
-        </div>
-
-        {/* CÉLULA 3: Adesão 7d (apenas em aba ativos) — S7.5.5: aplicar hover class */}
-        <div
-          className={`protocol-row-tabular__cell protocol-row-tabular__adherence-cell ${hoverClass}`}
-        >
-          {showAdherence && <AdherenceBar7d score={item.adherenceScore7d} />}
-        </div>
-
-        {/* CÉLULA 4: Estoque — S7.5.5: aplicar hover class */}
-        <div
-          className={`protocol-row-tabular__cell protocol-row-tabular__stock-cell ${hoverClass}`}
-        >
-          <StockPill status={item.stockStatus} daysRemaining={item.daysRemaining} />
-        </div>
-      </>
+      <ProtocolRowTabular
+        item={item}
+        showAdherence={showAdherence}
+        isHovered={isHovered}
+        onRowMouseEnter={onRowMouseEnter}
+        onRowMouseLeave={onRowMouseLeave}
+        onRowClick={onRowClick}
+        onDelete={onDelete}
+      />
     )
   }
 
-  // ────────────────────────────────────────────────────────────────
-  // VARIANTE: CARD (Modo Simples)
-  // ────────────────────────────────────────────────────────────────
-  const canExpand = isComplex && (item.hasTitration || item.notes)
-
-  function handleClick() {
-    // Se em modo complexo e tem conteúdo para expandir, expande
-    if (canExpand) {
-      onToggleExpand()
-    } else if (onEdit) {
-      // Caso contrário, abre o modal de edição
-      onEdit(item)
-    }
-  }
-
   return (
-    <div className="protocol-row">
-      {onDelete && (
-        <button
-          className="protocol-row__delete-btn"
-          onClick={(e) => {
-            e.stopPropagation()
-            onDelete(item)
-          }}
-          aria-label={`Excluir tratamento ${item.medicineName}`}
-          title="Excluir tratamento"
-        >
-          <Trash2 size={14} />
-        </button>
-      )}
-      <button
-        className="protocol-row__main"
-        onClick={handleClick}
-        aria-expanded={canExpand ? expanded : undefined}
-        style={{ minHeight: '3.5rem' }}
-      >
-        {/* Linha 1: nome + badge (esq) · stock pill (dir) */}
-        <div className="protocol-row__top-row">
-          <div className="protocol-row__header">
-            <span className="protocol-row__name">{item.medicineName}</span>
-            {item.concentrationLabel && (
-              <span className="protocol-row__dosage">{item.concentrationLabel}</span>
-            )}
-          </div>
-          <StockPill status={item.stockStatus} daysRemaining={item.daysRemaining} />
-        </div>
-
-        {/* Linha 2: quantidade por dose */}
-        <div className="protocol-row__intake">{item.intakeLabel}</div>
-
-        {/* Linha 3: frequência/horários (esq) · adesão (dir) */}
-        <div className="protocol-row__bottom-row">
-          <div className="protocol-row__schedule">
-            {item.frequencyLabel}
-            {item.timeSchedule.length > 0 && ` · ${item.timeSchedule.join(' / ')}`}
-          </div>
-          {showAdherence &&
-            (isComplex ? (
-              <AdherenceBar7d score={item.adherenceScore7d} />
-            ) : (
-              <AdherenceLabel score={item.adherenceScore7d} />
-            ))}
-        </div>
-
-        {canExpand && (
-          <span className="protocol-row__chevron" aria-hidden="true">
-            {expanded ? '▲' : '▼'}
-          </span>
-        )}
-      </button>
-
-      <AnimatePresence initial={false}>
-        {expanded && canExpand && (
-          <motion.div
-            key="expanded"
-            initial={{ height: 0, opacity: 0 }}
-            animate={{ height: 'auto', opacity: 1 }}
-            exit={{ height: 0, opacity: 0 }}
-            transition={{ duration: 0.22, ease: 'easeInOut' }}
-            className="protocol-row__details"
-          >
-            {item.hasTitration && <TitrationBadge summary={item.titrationSummary} />}
-            {item.notes && <p className="protocol-row__notes">{item.notes}</p>}
-          </motion.div>
-        )}
-      </AnimatePresence>
-    </div>
+    <ProtocolRowCard
+      item={item}
+      isComplex={isComplex}
+      expanded={expanded}
+      onToggleExpand={onToggleExpand}
+      onEdit={onEdit}
+      onDelete={onDelete}
+      showAdherence={showAdherence}
+    />
   )
 }
