+++ apps/web/src/views/redesign/Medicines.jsx
@@ -1,17 +1,15 @@
 import { useState, useMemo, useCallback } from 'react'
-import { Pill, Plus, CheckCircle2, AlertCircle } from 'lucide-react'
 import { useDashboard } from '@dashboard/hooks/useDashboardContext.jsx'
 import { medicineService } from '@shared/services'
 import Loading from '@shared/components/ui/Loading'
-import Modal from '@shared/components/ui/Modal'
-import EmptyState from '@shared/components/ui/EmptyState'
-import MedicineForm from '@medications/components/MedicineForm'
-import MedicineCardRedesign from '@medications/components/redesign/MedicineCardRedesign'
-import ConfirmDialog from '@shared/components/ui/ConfirmDialog'
-import './Medicines.css'
+import MedicineListHeader from '@medications/components/redesign/medicines/components/MedicineListHeader'
+import MedicineFilterChips from '@medications/components/redesign/medicines/components/MedicineFilterChips'
+import MedicineFeedbackBanners from '@medications/components/redesign/medicines/components/MedicineFeedbackBanners'
+import MedicineGrid from '@medications/components/redesign/medicines/components/MedicineGrid'
+import MedicineModals from '@medications/components/redesign/medicines/components/MedicineModals'
+import '@medications/components/redesign/Medicines.css'
 
 export default function Medicines({ onNavigateToProtocol }) {
-  // 1. Context & Memos
   const {
     medicines: contextMedicines,
     protocols,
@@ -20,7 +18,6 @@ export default function Medicines({ onNavigateToProtocol }) {
     refresh,
   } = useDashboard()
 
-  // 2. States
   const [error, setError] = useState(null)
   const [isModalOpen, setIsModalOpen] = useState(false)
   const [editingMedicine, setEditingMedicine] = useState(null)
@@ -30,11 +27,9 @@ export default function Medicines({ onNavigateToProtocol }) {
   const [showProtocolPrompt, setShowProtocolPrompt] = useState(false)
   const [newMedicineId, setNewMedicineId] = useState(null)
 
-  // 3. Calculate dependencies from dashboard context in-memory
   const medicineDependencies = useMemo(() => {
     const deps = {}
     if (!contextMedicines || !protocols || !stockSummary) return deps
-
     contextMedicines.forEach((med) => {
       const hasProtocols = protocols.some((p) => p.medicine_id === med.id)
       const hasStock = stockSummary.some((s) => s.medicine_id === med.id)
@@ -43,21 +38,9 @@ export default function Medicines({ onNavigateToProtocol }) {
     return deps
   }, [contextMedicines, protocols, stockSummary])
 
-  // 3. Feedback handlers
   const showSuccess = (msg) => {
     setSuccessMessage(msg)
     setTimeout(() => setSuccessMessage(''), 3000)
-  }
-
-  // 4. Action handlers
-  const handleAdd = () => {
-    setEditingMedicine(null)
-    setIsModalOpen(true)
-  }
-
-  const handleEdit = (medicine) => {
-    setEditingMedicine(medicine)
-    setIsModalOpen(true)
   }
 
   const handleSave = useCallback(
@@ -69,13 +52,12 @@ export default function Medicines({ onNavigateToProtocol }) {
         } else {
           const newMedicine = await medicineService.create(medicineData)
           showSuccess('Medicamento cadastrado com sucesso!')
-          // Show protocol prompt instead of window.confirm()
           setNewMedicineId(newMedicine.id)
           setShowProtocolPrompt(true)
           setIsModalOpen(false)
           setEditingMedicine(null)
           refresh({ force: true })
-          return // Don't close modal yet after refetch
+          return
         }
         setIsModalOpen(false)
         setEditingMedicine(null)
@@ -86,10 +68,6 @@ export default function Medicines({ onNavigateToProtocol }) {
     },
     [editingMedicine, refresh]
   )
-
-  const handleDeleteRequest = useCallback((medicine) => {
-    setDeleteTarget(medicine)
-  }, [])
 
   const handleDeleteConfirm = useCallback(async () => {
     if (!deleteTarget) return
@@ -106,20 +84,10 @@ export default function Medicines({ onNavigateToProtocol }) {
 
   const handleProtocolPromptConfirm = () => {
     setShowProtocolPrompt(false)
-    if (onNavigateToProtocol && newMedicineId) {
-      onNavigateToProtocol(newMedicineId)
-    }
+    if (onNavigateToProtocol && newMedicineId) onNavigateToProtocol(newMedicineId)
     setNewMedicineId(null)
   }
 
-  const handleProtocolPromptCancel = useCallback(() => {
-    setShowProtocolPrompt(false)
-    setNewMedicineId(null)
-    setIsModalOpen(false)
-    setEditingMedicine(null)
-  }, [])
-
-  // 5. Derived state
   const filteredMedicines = useMemo(
     () => contextMedicines?.filter((m) => filterType === 'all' || m.type === filterType) || [],
     [contextMedicines, filterType]
@@ -135,121 +103,24 @@ export default function Medicines({ onNavigateToProtocol }) {
 
   return (
     <div className="sr-medicines">
-      {/* Header */}
-      <div className="sr-medicines__header">
-        <div>
-          <div className="sr-medicines__title-group">
-            <div className="sr-medicines__title-icon">
-              <Pill size={22} />
-            </div>
-            <h2 className="sr-medicines__title">Medicamentos</h2>
-          </div>
-          <p className="sr-medicines__subtitle">Gerencie seus medicamentos cadastrados</p>
-        </div>
-        <button className="btn-primary" onClick={handleAdd}>
-          <Plus size={18} /> Adicionar
-        </button>
-      </div>
-
-      {/* Filter Chips */}
-      <div className="sr-medicines__filters">
-        {[
-          { key: 'all', label: 'Todos' },
-          { key: 'medicamento', label: 'Medicamentos' },
-          { key: 'suplemento', label: 'Suplementos' },
-        ].map(({ key, label }) => (
-          <button
-            key={key}
-            className={`sr-filter-chip ${filterType === key ? 'sr-filter-chip--active' : ''}`}
-            onClick={() => setFilterType(key)}
-          >
-            {label}
-          </button>
-        ))}
-      </div>
-
-      {/* Feedback Banners */}
-      {successMessage && (
-        <div className="sr-medicines__success">
-          <CheckCircle2 size={18} /> {successMessage}
-        </div>
-      )}
-      {error && (
-        <div className="sr-medicines__error">
-          <AlertCircle size={18} /> {error}
-        </div>
-      )}
-
-      {/* Content */}
-      {contextMedicines?.length === 0 ? (
-        <EmptyState
-          illustration="protocols"
-          title="Nenhum medicamento cadastrado"
-          description="Cadastre seus medicamentos para começar a controlar sua saúde"
-          ctaLabel="Cadastrar Medicamento"
-          onCtaClick={handleAdd}
-        />
-      ) : (
-        <div className="sr-medicines__grid">
-          {filteredMedicines.map((medicine) => (
-            <MedicineCardRedesign
-              key={medicine.id}
-              medicine={medicine}
-              onEdit={handleEdit}
-              onDelete={handleDeleteRequest}
-              hasDependencies={
-                medicineDependencies[medicine.id]?.hasProtocols ||
-                medicineDependencies[medicine.id]?.hasStock
-              }
-            />
-          ))}
-        </div>
-      )}
-
-      {/* Modal: Edit/Create */}
-      <Modal
-        isOpen={isModalOpen}
-        onClose={() => {
-          setIsModalOpen(false)
-          setEditingMedicine(null)
-        }}
-      >
-        <MedicineForm
-          medicine={editingMedicine}
-          onSave={handleSave}
-          onCancel={() => {
-            setIsModalOpen(false)
-            setEditingMedicine(null)
-          }}
-        />
-      </Modal>
-
-      {/* ConfirmDialog: Delete */}
-      <ConfirmDialog
-        isOpen={!!deleteTarget}
-        title={`Excluir "${deleteTarget?.name}"?`}
-        message={
-          medicineDependencies[deleteTarget?.id]?.hasProtocols ||
-          medicineDependencies[deleteTarget?.id]?.hasStock
-            ? 'Este medicamento possui tratamentos e/ou estoque associados. Esta ação não pode ser desfeita.'
-            : 'Esta ação não pode ser desfeita.'
-        }
-        confirmLabel="Excluir"
-        variant="danger"
-        onConfirm={handleDeleteConfirm}
-        onCancel={() => setDeleteTarget(null)}
+      <MedicineListHeader onAdd={() => { setEditingMedicine(null); setIsModalOpen(true); }} />
+      <MedicineFilterChips filterType={filterType} onFilterChange={setFilterType} />
+      <MedicineFeedbackBanners successMessage={successMessage} error={error} />
+      <MedicineGrid 
+        medicines={contextMedicines} 
+        filteredMedicines={filteredMedicines} 
+        onAdd={() => { setEditingMedicine(null); setIsModalOpen(true); }}
+        onEdit={(medicine) => { setEditingMedicine(medicine); setIsModalOpen(true); }}
+        onDelete={setDeleteTarget}
+        medicineDependencies={medicineDependencies}
       />
-
-      {/* ConfirmDialog: Create Protocol Prompt */}
-      <ConfirmDialog
-        isOpen={showProtocolPrompt}
-        title="Medicamento criado!"
-        message="Deseja criar um tratamento para ele agora?"
-        confirmLabel="Criar Tratamento"
-        cancelLabel="Depois"
-        variant="default"
-        onConfirm={handleProtocolPromptConfirm}
-        onCancel={handleProtocolPromptCancel}
+      <MedicineModals 
+        isModalOpen={isModalOpen} setIsModalOpen={setIsModalOpen}
+        editingMedicine={editingMedicine} setEditingMedicine={setEditingMedicine}
+        handleSave={handleSave} deleteTarget={deleteTarget} setDeleteTarget={setDeleteTarget}
+        handleDeleteConfirm={handleDeleteConfirm} medicineDependencies={medicineDependencies}
+        showProtocolPrompt={showProtocolPrompt} handleProtocolPromptConfirm={handleProtocolPromptConfirm}
+        handleProtocolPromptCancel={() => { setShowProtocolPrompt(false); setNewMedicineId(null); setIsModalOpen(false); setEditingMedicine(null); }}
       />
     </div>
   )
